integrate autosuggest onto testrequest component
