+++ stockTrainerFrontend/src/App.js
@@ -16,7 +16,6 @@ import Help from "./components/help/Help";
 
 import Callback from "./Auth/Callback";
 import TestRequest from "./components/TestRequest";
-import TestSearch from "./components/TestSearch";
 
 class App extends Component {
   constructor(props) {
@@ -190,16 +189,6 @@ class App extends Component {
                   retrieveStock={this.retrieveStock}
                   stockData={stockData}
                 />
-              </div>
-            )}
-          />
-          <Route
-            exact
-            path="/testsearch"
-            render={props => (
-              <div className="lowerPageLayout">
-                <NavBar {...props} />
-                <TestSearch />
               </div>
             )}
           />
